allow the max number of concurrent db connections to be controlled
by default database/sql doesn't cap the number of open connections, so a
high traffic pgredis server could generate very high load on the
database.

It's worth giving the pgredis administrator a way to manage the amount
of concurrent traffic that's sent to the database. In many cases it's
probably preferable for the traffic bursts to be queued in pgredis
rather than the database.

Closes #11