WIP: kinda working, but hit a roadblock
I forced this to a point where I can use MULTI and EXEC to apply
multiple commands, and extracting db transaction management to the top
level seems like it could work.

However, the requirement for EXEC to return an array of responses for
each individual command is a roadblock.

At the moment each command serialises it's response directly to a buffer
and it's not straight forward to reformat them into an array.

More thinking required.