remove resposibility for protocol serialisation from the command interface
Until now, each command implementation has been responsibile for
serialising it's response. That's worked ok, but it's not very flexible.
For example, there's no way for the connection control loop to examime,
store or buffer the command response.

This creates a new set of types that confirm to the pgRedisValue
interface. pgRedisValue defines a writeTo(io.Writer) function that will
serialise the type using the redis protocol.

Every command must now return a pgRedisValue instance, and the
connection handling code can do with that what it wants. For now, that's
just serialising it directly to the client. From the clients
perspective, no change in behaviour for now.