partially implement SSCAN
This implementation is NOT usable in the real world, but it is
technically correct.

Just to get things going, we're always returning the full set. The
returned cursor is 0 so hopefully clients will respect the response.
However, in the real world sets could be very large and SSCAN is
explicitly designed to avoid serialising the full set. Clients will be
using it to iterate over large sets, and they'll be surprised if we send
everything

Contributes to #26