switch to my fork of go-redisproto
The upstream version allows a maximum of 20 arguments in a single
command, but the sidekiq test suite has some cases that send commands
with up to 1002 arguments.

I've forked the lib and increased the maximum to 1024 arguments. If it
works well, I might open an upstream PR to make the maximum configurable.