Add stubbed implementations of CLIENT, SELECT and FLUSHDB
These are all used by the sidekiq test suite so I've added stubbed
implementations.

SELECT: selects the database to work in. We don't support multiple
databases (yet?), so just always return "OK". This could be risky and
it'd be easy for a user to think they've changed database before
destroying data. Must revisit soon!

FLUSHDB: like FLUSHALL, but only for a single database. We don't support
multiple databases yet, so just handle this with the FLUSHALL code

CLIENT SETNAME: allow the client to name their connection. For now, we
have nowhere to store the name so we throw it away and tell the client
that's "OK"