trial using advisory locks to avoid concurrency issues
I'm getting intermittent spec failures in RPUSH and LPUSH.

I'm still trying to get to the bottom of the failures, but they seem
very racy and I strongly suspect concurrency issues (which is weird,
because the test suite isn't concurrent....).

Maybe I need to consider using advisory locks before making any changes
to the database.... just to be 100% sure that only one connection is
changing a key at a time?

Presumably this will hurt performance. Also, if it helps then maybe we
need to do this consistently for all commands that mutate the DB?