Refactor the connection handling loop to be more readable
This was originally implemented as a hacked together state machine, built
as I got the MULTI specs green. I successfully got to a passing build,
but always wanted to refactor it into something more readable.

Here's my first attempt at that. I've dropped the state machine and gone
with an in-memory request queue instead. After each request is added to
the queue, we examine the queue state and determine if any action is
required.

I'm not totally happy with the result, but it's better than what I had
so I'm going to merge it.

This approach has a bonus side effect: delaying the DB transaction until
a `MULTI` is ready to `EXEC` means that we will have the option to
select **all* keys that will be modified in the DB transaction and lock
them in alphabetical order to avoid deadlocks.

No specs were changed, and the suite is still green.

Closes #44