move all psql locks to the  start of the transaction
To guarantee we get no deadlocks, we need to take **all* the key locks
in alphabetical order and before we start executing any of the commands
that modify the database.

This moves all database lock acquisition to the start of the
transaction, and removes it from all of the repository methods. The
repositories now assume it's safe to just jump in and make their
changes.

There's still a small inefficiency: we don't remove dupe keys before
acquiring locks. This should have no user visible impact, other than a
very small performance hit.

I've also avoided taking any locks on BRPOP, even though it modifies a
key. In the new structure we don't want to take the lock for BRPOP at
the start of the transaction, because that DB transaction will be open
for a long time and we don't want to block other requests. This needs to
be fixed, but I'll open a separate GH issue.

Fixes #48